+++ libs/base/control.cpp
@@ -1,4 +1,4 @@
-#include "pxt.h"
+#include "pxtbase.h"
 
 /**
  * How to create the event.
