+++ libs/base/core.cpp
@@ -1,4 +1,4 @@
-#include "pxt.h"
+#include "pxtbase.h"
 #include <limits.h>
 
 namespace String_ {
