+++ libs/base/loops.cpp
@@ -1,4 +1,4 @@
-#include "pxt.h"
+#include "pxtbase.h"
 
 namespace loops {    
     void forever_stub(void *a) {
