+++ libs/base/pxt.json
@@ -6,7 +6,7 @@
         "dal.d.ts",
         "pxt-core.d.ts",
         "pxt.cpp",
-        "pxt.h",
+        "pxtbase.h",
         "core.cpp",
         "pxt-helpers.ts",
         "buffer.cpp",
@@ -23,6 +23,7 @@
     "testFiles": [
         "test.ts"
     ],
+    "additionalFilePath": "../../node_modules/pxt-common-packages/libs/base",
     "public": true,
     "dependencies": {}
 }
