+++ libs/base/serial.cpp
@@ -1,4 +1,4 @@
-#include "pxt.h"
+#include "pxtbase.h"
 
 //% weight=2 color=#002050 icon="\uf287"
 //% advanced=true
