+++ libs/core/pxt.json
@@ -4,27 +4,17 @@
     "files": [
         "README.md",
         "dal.d.ts",
-        "pxt-core.d.ts",
-        "pxt.cpp",
+        "codal.cpp",
         "pxt.h",
         "pins.h",
-        "core.cpp",
         "pins.cpp",
         "devpins.cpp",
         "devpins.h",
         "pins.ts",
         "i2c.cpp",
         "i2c.ts",
-        "pxt-helpers.ts",
-        "buffer.cpp",
         "shims.d.ts",
         "enums.d.ts",
-        "control.cpp",
-        "control.ts",
-        "loops.cpp",
-        "serial.cpp",
-        "serial.ts",
-        "math.ts",
         "hf2.cpp",
         "hf2.h",
         "uf2format.h",
@@ -45,6 +35,9 @@
             }
         }
     },
+    "additionalFilePath": "../../node_modules/pxt-common-packages/libs/core",
     "public": true,
-    "dependencies": {}
+    "dependencies": {
+        "core": "file:../base"
+    }
 }
